Separated the modulation code into its own namespaces and modules
